Added environment variable parsing and formatting routines.
The default environment variable API in the os package is a bit lacking.
In particular, it doesn't provide a nice way to iterate over environment
variables or copy the environment. This package adds functions for
parsing environment blocks (sequences of key=value strings) and
converting them to a more sensible map data structure. It also adds
functions for formatting maps back into environment blocks, which will
be necessary for setting subprocess environments.