Added build script and empty stubs for the CLI and agent.
The build script generates agent bundles and release bundles containing
CLI binaries and agent bundles. It can be run in a development mode that
will only generate a CLI executable for the current platform and a
limited agent bundle for a set of common testing platforms.

The build script doesn't use upx on agents at the moment, because the
vast majority of our platforms aren't supported by UPX. Instead, it just
gzips the agent bundles, which will have about the same effect on total
size, but will make agent extraction and transfer slower than if we
could simply UPX and tar.

Unfortunately, because file monitoring is going to require CGO on
Darwin, the build script will warn when building Darwin agents on
non-Darwin platforms, and flat out won't allow release builds on
non-Darwin platforms. As long as we don't have any more CGO dependencies
arise on other platforms, we should be okay to just have Travis upload
release bundles from the Mac builder.