Added file locking facilities.
This commit adds a filesystem package with a locker type that allows for
cooperative locking on the filesystem. On POSIX this uses FcntlFlock and
on Windows this uses LockFileEx.

Unfortunately Plan 9 doesn't have any sort of file locking API, so we
have to exclude it for now.

This locking has the important property that if the process holding the
lock dies without releasing it, the OS will automatically release the
lock. On POSIX systems, this should happen almost instantly. On Windows
systems, this apparently only happens after some unspecified period of
time, though it appears to be effectively instant.

This locking suffers from the standard fcntl lock quirks, namely that if
the process closes ANY file descriptor pointing to the locked file, the
lock is released. But this shouldn't be a problem in practice since we
have at most one reference to the daemon lock per process.

How this locking behaves when forking is not really consistent between
platforms, so it's best not to hold a lock across a fork.