Added filesystem case sensitivty and Unicode decomposition testing.
This commit adds tests for case sensitivity and Unicode decomposition.
Both use the same underlying method to check if the filesystem is
sensitive to variations when accessing temporary files.

Case sensitivity is obviously important to consider when dealing with
multiple platforms. The Unicode decomposition problem is
Darwin-specific, though we actually test for it instead of just looking
at the OS because it also depends on the filesystem (and Apple is coming
out with a new one in the next year or so).

We didn't add a test for executability bit propagation, because there's
not really any test (other than OS) that would be useful in determining
how executability for filesystem entries should be handled.