Vendored the github.com/gogo/protobuf package.
This package provides accelerated serialization routines for Protocol
Buffers messages. It is actually a fork of the
github.com/golang/protobuf packages, but adds a number of extensions for
things like stable serialization (which we need). Unlike the nominal
protobuf package, which generates some code but relies heavily on
reflection to perform serialization, this package generates code for
serializing each message type, generally resulting in much better
performance and lower memory usage. It is used by almost all serious Go
projects doing Protocol Buffers serialization and is compatible with
gRPC.

This package does provide tags, but the most currently release is vastly
out of date and has no proto3 support. HEAD does, however, so we've left
the package at HEAD.

There will be some juggling necessary in order to keep the vendored
protobuf, gRPC, and gogo-protobuf packages at compatible points, and
while things have largely settled down in this arena, it's essential
that these be considered together when updating submodules.