Vendored the gopkg.in/natefinch/npipe.v2 package.
This one is a bit tricky, because this uses gopkg.in to redirect to a
branch on GitHub (v2). This is the default branch, so it's what's been
checked out, but we'll have to be a little careful when updating this
dependency by making sure we're always checking out the latest commit on
the correct branch.