Added daemon IPC infrastructure.
On POSIX, we use UNIX domain sockets, while on Windows we used named
pipes. Both of these provide authentication and authorization at the
connection level, so we don't need to perform additional authentication
on top.

On Windows, we have to generate a random named pipe name and write it to
a secure location to stop squatting attacks and superfluous pipe
connectivity. Named pipes are really dangerous due to their default
impersonation abilities, but the approach we take should address these
problems.

Named pipes are kind of an ugly beast, and the Go implementations of
them are even uglier, but they're sufficient for now. In the future, we
may want to consider switching to Microsoft's go-winio package
(https://github.com/Microsoft/go-winio), but unfortunately it's a bit
nascent at the moment. It provides saner defaults though for
impersonation settings, allowing authentication without impersonation.