Modified session creationTime field to use gogoproto.stdtime.
Apparently gogo/protobuf's generated code doesn't work without this.
It's not a big deal, since it's actually a convenience, but I'm starting
to grow weary of the gogo/protobuf situation. It'd be really nice if we
could just avoid using it (at the cost of a small amount of speed) and
have deterministic serialization built into the default Go protobuf
package like it's built into the C++ Protocol Buffers implementation.