Added the vendor path before $GOPATH/src in Protcol Buffers import path.
This is necessary so that protobuf and gogo/protobuf imports are pulled
from the vendor directory. This doesn't really affect Well-Known Types
because they're pulled from the protoc installation anyway, but the
gogo/protobuf extensions are imported and need to match the runtime that
we're going to compile against.

Note that the WKTs will have their import path set to
github.com/golang/protobuf since the protoc installation proto files
define the go_package option, but since we have golang/protobuf
vendored, the implementations will be found in the vendor director when
compiler. This commit has no bearing on this fact.