Added URL classification and parsing facilities.
The new url package supports classification and parsing of URLs. It
currently only supports local paths and SCP-style URLs. It uses some
heuristics in classification and parsing that are non-standard (e.g.
correctly identifying Windows paths and supporting ports in SCP URLs),
but these should be pretty robust. The package should be easily
extendable to other formats and protocols.