Extracted Endpoint and Prompt service and messages into agent package.
I'm a little bit apprehensive about separating the session
synchronization logic between session and agent, but since the endpoint
lives in the agent, it might be okay. Otherwise, we might want to move
the synchronization loop and endpoint into the sync package, and make
session and agent depend on that. Right now I think session should
depend on agent, but we can't have agent depend on session. But
actually, it doesn't need to, because the Endpoint only needs to live in
the agent binary. I'll fix this in the next commit.