Moved Endpoint back into session.
This is fine, because there won't be a ciruclar dependency (Endpoint
instances will live in the agent binary, and won't be used in the agent
package). All the agent package needs to know is how to speak gRPC. Of
course, it'd be more natural if the agent package could return an
endpoint client instead of a client connection, but I can live with
this.