Disabled Close on stdioConn.
Unfortunately there's no way to faithfully implement the semantics of
net.Conn.Close since calling Close on standard input/output can block if
they are being read/written to. But it doesn't matter since the
connection lifetime is tied to the lifetime of the process, so the
process just exits.