Major refactoring of agent connectivity and installation.
This commit takes all SSH functionality out of the agent package and
moves it into a dedicated ssh package. It also removes the agent service
and makes all of its methods stateless. It cleans up the SSH
implementation and divides up the agent package into components that
will lend themselves better towards extensibility to other protocols.
Prompting has also been refactored so that it takes place in a side
channel.

At this point, the agent and SSH code is pretty lean and mean, probably
best to start on synchronization.