Removed circular dependency between agent and session.
Well, technically this dependency didn't exist yet, but it would have
once the agent package tried to start registering endpoints with its
gRPC servers. To fix this, we've had to move that into the agent process
(thus creating a compilation boundary where the only dependency is
protocol compatibility) and move local endpoint creation into the
session package. I don't LOVE this latter change, but it kinda makes
sense in a way, and would be more natural if we eventually moved
Endpoint to become an interface that is perhaps satisfied by gRPC for
remotes. But for now, this is okay - agent takes care of remote
endpoints, as it was always supposed to, and local ones are simple
enough to create in-memory.

The only downside is that we might run into trouble if we want to rely
on the agent for additional services. But actually, in a way, by moving
service registration into the relevant body, things become more
sensible. We could, for example, register a filesystem service with the
agent gRPC server that we don't register (because it isn't needed) with
the in-memory server.