Moved server creation and service registration into daemon entry point.
This feels a bit more natural and more in-line with what is done for
agent. Encapsulation is nice, but we already have other daemon
components that aren't fully encapsulated, and as we add more things
(perhaps even custom services for other daemon host processes), it just
makes more sense to have these services more accessible.