Removed resolution infrastructure and clean up reconciliation logic.
This commit removes conflict resolution infrastructure. Implementing
resolutions was going to be very complicated from a technical
perspective. E.g., we would need to communicate with the synchronization
runloop and inject either a resolution dictionary or callback, neither
of which would have been very feasible. Additionally, we would only be
able to potentially resolve both-modified conflicts, and the resolution
would always involve deleting the contents of one side. Given this, it
makes more sense to have the user just delete the side that they don't
want on disk.

This commit also cleans up the reconciliation logic and adds better
comments to illuminate its functionality.