Added DirectoryContents routine to filesystem package.
This routine will encapsulate any Unicode normalization that needs to
occur (e.g. on HFS). It is also name-only, because for our use case it
is better to stat just before an operation (and sometimes we only need
names) (otherwise we'd just use io/ioutil.ReadDir). This routine also
uses statfs to avoid having to do the round-tripping test (which might
not work on a read-only filesystem and might not be correct if there is
a mount mid-way through a tree).