Added context support to SSH, integrated with agent.
This will allow us to more easily cancel SSH connection attempts, which
will be useful for the session service where we'll use contexts to
terminate synchronization loops.

We've used golang.org/x/net/context for now, because although it is
interface-compatible with the standard library context package, it uses
a lot of internal casts to do its magic (pretty much entirely negating
the purpose of having Context be an interface). Once gRPC switches over,
we'll do the same. It's worth noting that we implement this SSH context
support using os/exec.CommandContext, which DOES take a standard library
context, but in that case it's only using the actual Context interface
(specifically its Done method), so there's no internal magic there that
we have to worry about, and it doesn't look like that will change
because there's no need for it to in that particular case.

This commit almost certainly won't build, but it is part of a major
session refactor effort, so you'll have to look ahead.