Removed context usage in agent dialing.
This didn't really jive with the semantics of context - the context
should cancel the operation that computes a result, not the result
itself. Unfortunately with the SSH case, there was no way to
satisfactorally associate a context with dialing, because the context
would have become attached to the command. If we were using the
synchronization loop context for this, it would be fine, but when doing
start or resume, we would have wanted to use the handler context to
cancel dials, and we don't want the resulting SSH processes to be scoped
to the handler context or they'd die instantly. Instead, we'll just use
the context at a higher level, ignoring any dial results if they come
too late. This won't cause any problem since agent connections will just
die off automatically. This is similar to what gRPC does with its
timeout.