Added local dialing, unified dialing routines, added context to dial.
This commit modifies the agent package to support local dialing. We're
moving local agents out-of-process. The exact mechanism is
well-abstracted, so we can probably find a way to work around this later
if we need to bring local agents back in-process (though it'll involve
working around a circular dependency between the session and agent
packages).

This commit also modifies agent dialing into a single Dial function that
just takes the new URL type.

Finally, this commit adds support for cancelling dials via a context.