Added a URL Protocol Buffers message type, refactored URL parsing.
Originally I thought that keeping URLs as strings in sessions would be
more flexible, but I realize now that it is actually less flexible
because it locks in URL parsing semantics for-ev-er. By adding a URL
type that has fixed but extensible semantics, we can freely vary the
definition and format of URL strings later on (though we probably
won't). In any case, it allows us to change the parsing code without
fear that we'll break string URLs in existing sessions.