Adjusted agent and ssh packages for new URL type.
Fortunately this was simple, all we had to do was change the enumeration
names. I'm still not sure if we want to use the validation method in the
URL, because I feel like it adds a lot of knowledge in the URL package
about how the URLs will be used, but we need the Start handler to be
able to validate URLs without having to reach into the SSH or agent
package (because then it knows too much). But maybe it just needs to
validate that the path is non-empty? I'm not sure just yet.