Massive refactor and work on implementation of session package.
The changes here are too numerous to detail, but broadly speaking the
bulk of the session package has been implemented. It is still lacking
the actually synchronization loop and a few endpoint methods, and there
are some bits here and there that are missing. But the mutagen binary
now compiles.

It goes without saying that ALL of this code is untested.