Updated build script with a more flexible set of build modes.
There is now a 3-tier hierarchy of build modes: slim, testing, and
release. Each builds for a broader range of targets. This is designed to
make development compilation a bit simpler and faster.

I've also disabled the Plan 9 build completely, even for agents, because
it is becoming unmanagable. Once we have a release candidate, we can
play with getting Plan 9 to build, but right now the source is still in
flux a bit too much to commit to the complex build tag structure that
it's going to take to make Plan 9 work.