Disabled the requirement that file contents be non-empty for decoding.
An empty byte slice is technically a valid Protocol Buffers message.
The JSON unmarhsal method does NOT like an empty input message, but the
json.Unmarshal function will tell the caller that, so it's better to
just push that message higher up, and it won't be a problem for us in
practice.