Added no-frills RPC package with the hopes of removing gRPC.
This package only allows a single method invocation per connection, but
this is fine for the command line client and will be set on top of a
stream multiplexer for controller-endpoint connections. It uses object
streams built on top of the encoding/gob package. This allows for
streaming RPC with dynamic typing. It's a heck of a lot smaller than
gRPC, a heck of a lot more flexible, and I suspect a heck of a lot more
performant since there's not a whole HTTP/2 stack to go through.