Modified command handlers to return errors instead of aborting.
This change allows handlers to return errors and have them trigger and
abort instead of having to trigger the abort themselves. This allows the
handlers to more easily defer cleanup behavior and be sure that it will
execute. Handlers are still permitted to exit the process on their own
if they want to (e.g. in argument parsing where a help flag will
naturally trigger an exit), but this doesn't hamper their ability to do
cleanup.