Gutted the session package and created stubs for endpoint.
This commit guts out most of the existing gRPC-based session,
controller, and endpoint logic. It replaces these with method stubs. It
adds an endpoint serving function that can be used to run an endpoint
server on a given stream. This will be used for in-memory endpoints as
well.

This commit also performs a bit of cleanup in unrelated code in order to
keep symmetry with modifications.