Added endpoint connection function to session package.
This necessitated a few changes to the agent package and binary, but
will give us a more flexible and extensible mode defintion mechanism.

For local endpoints, we use an in-memory server, making this about as
close as possible to a zero-cost abstraction (though we still incur an
in-memory copy of all streams).