Refactored sync package to use non-map-based Entry type.
This necessitated touching every part of the sync package. Other
packages likely won't build at the moment. I'm going to move back to the
rsync-based snapshot transfer, and this is the easiest and most
performant way of doing it.

The transition code really needs to have a once-over, as there are some
asymmetries between the removal and creation routines.

I've also refactored the scan infrastructure to look a bit more like the
rest of the code in terms of path handling.

Ignores still aren't implemented, but should be soon.