Added ignore support.
This uses the doublestar package to implement its ignores, with
additional support for pattern negation. The performance is acceptable
for now. For the Linux repo, each ignore pattern seems to add about 40ms
to the warm scan time, so maybe we can look into switching to a more
optimized mechanism later (see github.com/docker/docker/pkg/fileutils),
but I think we're okay for now.