Added PropagateExecutability to propagate executability bits.
This function will analyze an ancestor and propagate executability to
snapshots from systems that don't preserve executability (namely
Windows).

This function will mutate the snapshot in-place rather than making a
copy of it, in slight contrast to our existing convention. But I think
we'll change that convention and simply make it against policy to mutate
an entry after using it in Diff or Reconcile. Actually, since we'll
likely make Diff internal, mostly just for Reconcile. Effectively,
Reconcile will "take ownership" of the entries passed to it (though Go
provides no way to enforce this, so it'll have to be convention).

This solution will handle most cases, though it won't take care of POSIX
systems with filesystems that don't propagate executability.
Fortunately, the only case of this I can think of is FAT32 on POSIX, but
I think the overlap between FAT32 usage and Mutagen usage will be quite
low.