Added/expanded the staging interface in the sync package.
I still want to keep the sync package agnostic of the exact staging
mechanism, partly for future changes and partly because it's already a
big package. That being said, the sync package does need to know about
staging, and it can implement a few utility methods that don't tie us to
a particular staging strategy, so I've added those.