Changed executability functions to mutate copies, added stripping.
This commit modifies PropagateExecutability to operate on a copy of the
snapshot (like Apply and Transform). This particular algorithm could
probably be remade to only allocate on changes, but it'll take some
thought (it's hard to allocate a contents slice if only some element in
the slice changes, but should be possible with index tracking and a
copy). Anyway, copy operations aren't a huge time driver so we'll take
this simple approach for now.

This commit also adds a StripExecutability function that'll be necessary
to compute expected snapshots for systems that don't support
executability propagation.