Changed how rpc services register their methods.
This allows for some inversion of control with regard to method
registration. Now services are registered individually via a simple
interface. This is closer in design to net/rpc and gRPC, but it doesn't
rely on reflection, just a simple method that each service neesd to
implement. This allows much more flexibiliy and encapsulation than
registering methods in the rpc server constructor.