Added food for though to the CLI daemon entry point.
I think we may eventually want to relocate some of this daemon setup
logic, but I'm not sure the right boundary to use for encapsulation and
I want to keep things extensible. Anyway, I don't want to forget about
this idea, because I think it's important to consider in detail, both
its pros and cons, for later integration into other processes.