Unexported diff function, changed how Conflicts are represented.
This change to Conflict is simple (removing the "root" path of the
conflict and making all conflict change paths "absolue" relative to the
synchronization root), but it jives much better with how we plan to
actually display conflicts. Trying to present users with the idea of the
"root" of the conflict is a bit abstract - it's better just to show them
which sets of changes are conflicting, and it's better to show them by
giving them full paths. There was no way we were going to present full
trees when showing conflicts anyway.

We still need to add a better stringer implementation for Change, but
even what we have now is probably sufficient for display.