Refactored Entry to use a content map AND have consistent serialization.
You read that right, and this is a bit of a hack, but a supported one.
Essentially, Protocol Buffers maps are encoded exactly the same way that
we were encoding our NamedEntry types - i.e. repeated 2-element
messages. So we create a special entry type (OrderedEntry) that is
equivalent to Entry but which serializes consistently, and then we just
perform a conversion and sort right before marshalling. This lets us use
maps in-memory (hooray!, although we still have to allocate to iterate,
which isn't great, but also isn't a performance nightmare), as well as
have consistent serialization. Even better, it allows us to avoid the
need for "gogo" Protocol Buffers (which would bloat binaries), while
still being completely supported and future-proof.