Added EnsureValid function to entry and added ancestor validation.
Protocol Buffers message types are unfortunately rather limited when it
comes to enforcing any sort of invariants, and this is exacerbated by
the needs of Go (and the Protocol Buffers Go implementation) where
reflection requires exporting fields. There's no good way to get around
this, particularly if we're going to rely on gob encoding for things
that contain entries. We MIGHT be able to get around it a bit by
encapsulating root entries in the sync package, but entries are also
exposed by change objects, so that doesn't work.

Thus, we're stuck manually validating that entry messages are correct.
In order to facilitate this, this commit adds an EnsureValid function
that enforces the entry invariants we care about. We also invoke it just
before saving the ancestor, because that's the most important place to
enforce correct entry formatting.

It might be worth adding this check elsewhere (e.g. when decoding
snapshots from remotes), but it's not that important then. It's really
just a check of our internal logic, which unfortunately can't be done by
the compiler.