Replaced filesystem monitoring package on Darwin.
The github.com/rjeczalik/notify package wasn't quite cutting the
mustard, so I've replaced it with the github.com/fsnotify/fsevents
package that's being considered for integration with the
github.com/fsnotify/fsnotify package. Unlike the fsnotify package, the
fsevents package supports recursive watches (part of the reason it
hasn't been merged yet). We still use the notify package on Windows, but
on Darwin it just has too many panics that we can't control (and
probably wouldn't want to even if we could because they are very
frequent and probably leaking resources).