Moved and refactored stable snapshot serialization and transfer.
This stable serialization is only necessary for our rsync-based snapshot
transfer strategy, so it belongs in the session package.

Furthermore, the previous implementation had a major bug - a root
directory entry serialized to an empty byte slice, so there was no way
to differentiate between that and a nil root. This was causing
reconciliation conflicts because the deserialization snapshot looked
nil.

To fix this, we've wrapped up the stable entry in a stable archive.

I've also fixed associated tests and scan_bench, though I think we need
to get benchmarking for this strategy done at some point.