Extracted path computation from endpoint. Made staging use prefixes.
This change extracts cache and staging path computation from the
endpoint and into standalone functions.

It also modifies the staging layout to use prefix directories based on
the first byte of the digest. For large numbers of files (say tens of
thousands) this will distribute the load of their storage over 256
nodes, making lookup a bit faster.