Refactored RPC streams to carry errors, other minor changes.
This commit refactors RPC streams to carry errors in headers that frame
messages. This allows RPC handlers to return errors and have them
received by clients.

Because of these new semantics, some methods have been refactored on the
client and server, namely stage, create, and resume.

We still don't really enforce stream semantics - clients and servers are
able to send as many messages as they like of any type, we just provide
a sideband for errors and a mechanism for indicating completion
(io.EOF).

I've modified create and resume to use the same prompter type. I'll soon
extract the code from the client side of create and make it shared.

I also took out some error cases in rsync that can't occur.

This is a big change, but makes handler code a lot cleaner and safer.