Fixed major transition bug.
It turns out that errors.Wrap(nil, ...) will return nil! This is a major
footgun in my opinion, but oh well. Anyway, it was causing an error to
be hidden during transition, meaning that a removal was marked as
successful when it actually wasn't. This was causing removals to be
propagated to the ancestor when they weren't complete, so on the next
synchronization cycle they just reversed themselves.