Added new rsync implementation.
The bitbucket.org/kardianos/rsync implementation was a little bit long
in the tooth (w.r.t. modern Go versions), had a very serious bug where
targets with a length that was a multiple of the block size weren't sent
completely, was very hard to read, and did a lot of allocating. It also
had some weird aspects, like the signature writer, taking operations via
a channel (which required a Goroutine for each transmission stream in
our use case), and having some questionable code.

Our new implementation has fewer knobs, most notably it fixes the strong
hash and doesn't allow for a configurable maximum data size or block
size. This allows us to avoid a lot of allocation. That being said,
adding these knobs in would not be hard - we'd just need to move from
free functions like we have to a type like the kardianos package has,
and probably check for a few invariants.

This implementation is working and faster, though it needs a test suite
added. It also doesn't use all the latest and greatest theoretical
aspects of rsync (it's heavily based on the original thesis and
technical report), but it is fine for our purposes.

I'm hoping to simplify the Deltafy function a bit.