Relocated stream interfaces, forwarding, and multiplexing facilities.
The Opener and Acceptor interfaces, as well as the Forward function, are
used exclusively for RPC, and I think fit better in that package.

The multiplexing facilities fit well in the stream package, so I've
relocated them there to help balance the size of the session package.

Other code has been refactored to adjust.