Added TODO item for adding version constraints in Go 1.8.
We're only going to support the latest version of Go, and a build
constraint in the version file is the best way to do this. We require
version 1.7.3+ at the moment, but there's no way to express that, so
we'll just wait for Go 1.8 before we start adding hard requirements.