Added a backpressure mechanism to regulate monitor responses.
This commit adds a backpressure mechanism (an empty "ready" request) to
monitoring list RPC calls. This keeps the daemon from sending more
updates than the client can process. Although the IPC buffer filling
also would have regulated this, this has the additional benefit of
reducing the number of state change polls that need to occur on the
server (while still not missing any).