Removed default ignores functionality.
These may come back in the future, but at the moment they are somewhat
at odds with the simple-but-configurable philosophy that Mutagen tries
to follow.

If we do bring these back, perhaps we want to divide them up into
different categories, e.g. system ignores (.DS_Store, etc), SCM ignores
(.git, .svn, .hg, etc), editor ignores (*.swp, etc), and so on. Or
perhaps we want to make a global ~/.mutagen/config where this could be
specified for the system as a whole (particularly for system ignores).

In any case, there's a LOT to think about.

But for now, users can just manually specify ignores and get behavior
that is much easier to reason about.