Refactored transition to continue in the case of problems.
This commit allows transitions to continue if they encounter problems,
and to report multiple problems for a single transition.

Previously transitions would stop on the first error, but now they'll
continue on to create as much content as possible. This was necessary
for a lot of reasons. One of the most important reasons is that the
endpoint always wipes the staging directory after transition, so if
transition failed after a single file, then tons of staged files might
be lost and then retransfered... EVERY TIME.