Disabled executability setting on agent binary extraction.
Windows doesn't support doing this, and for SSH it has to be done as part
of the install step anyway, so just take this code out for now. If we
decide to do out-of-process local agents again, we'll need to find a way
to do this inside the local connection routine. But we just don't need it
for now.