Made monitor status bar one column narrower each.
The cmd.exe console width is 80 characters by default. If output is
exactly 80 charcters and then "erased" with a '\r' character, then the
console will just wrap to the next line. To work around this, make the
columns a bit narrower for now. It's unlikely anybody will purposely set
their consoles narrower.

In the future, it might be better to switch to using some sort of
curses-like interface to watch the console for width changes and set size
accordingly.