Came up with a better solution for temporary scan failures.
Essentially we just add a TryAgain flag to scanResponse, but this allows
us to continue without reconnection. If we see that there are other
failure modes in the future that can occur like this, maybe we need to
consider modifying the RPC framework to delineate transport/remote
errors, or some other magic.

This solution is nice though because it allows the handling of these
failures to occur on the controller, doesn't require modification to the
RPC framework, and gives the user better feedback on what is occurring.