Disabled recursive watching using notify package on Windows.
Unfortunately the notify package has a fairly major bug where watches
can't be re-established on Windows, and it seems to apply even when
watches are cleanly shut down. We're probably going to have to just ditch
that package and write our own small ReadDirectoryChangesW wrapper. But
that's not the worst thing - it'll help out Windows binary size a bit.

The bug, for reference:

https://github.com/rjeczalik/notify/issues/69

Unfortunately, although I did know Windows watching was flakey, I didn't
catch this until just now because that issue was marked as closed. But all
we have to do is fall back to polling on Windows, which isn't too bad.
This change will at least shrink the polling Window a bit, making things a
bit more tolerable.