Enabled release bundle upload from Travis CI.
This necessitated adding a little conversion script to convert Windows
release bundles to zip format. I don't mind this little hack for now,
and I probably won't try to recode it in the build script, because
adding support for zip in Go seems to be a pain, especially in tandem
with tar.