Added a build constraint to enforce Go 1.8+ for compilation.
We haven't actually made any changes that require Go 1.8 yet, but this
branch will be full of them. In general, we'll only support the latest
version since there's really no difficulty in using the latest. Yes, it
will mean dropping support for some older platforms, but these don't
seem to be well-supported in older Go versions anyway.

I'm not a huge fan of our build constraint mechanism here, because it
doesn't really provide any feedback as to why Mutagen is failing to
build, and it relies on code trying to pull in Mutagen version
information, but there's not really any other way to approach this.