Added rename that can copy across FS boundaries, used in transitions.
The standard os.Rename function doesn't support renaming across
filesystem boundaries. This commit adds a wrapper function that can fall
back to a copy-rename-remove-based rename that still has the atomic
properties we want (i.e. it still uses a rename to move data into
place). It only supports files, not directories, but that's okay because
we create directories in place, we only need this function to rename
staged files.

This commit also updates the transition code to use this new rename
function.