Fixed Travis deployment conditionals (hopefully).
What I've done is just combine both conditions into a single
condition. The Travis documentation is super vague, gives inconsistent
examples, and there are a ton of open and historical Travis issues
related to deployment conditions that make me think we're best off
sticking with simple and stupid.

According to the Travis documentation, the conditions will be passed
into a Bash conditional of the form:

    if [[ <condition> ]]; then <deploy>; fi

See:

    https://docs.travis-ci.com/user/deployment#Conditional-Releases-with-on%3A

It also says that ALL conditions have to be met in order for deployment
to happen, so I assume it's a bit more complicated than that, but with
only one custom conditional, it should be pretty fool-proof.

I suspect that Travis is actually NOT handling multiple custom
conditions correctly. I think it is ORing them rather than ANDing them,
based on builds I had where Linux happily tried to deploy with the
following condition present:

    $TRAVIS_OS_NAME = osx

Fingers crossed!