Vendored the golang.org/x/sys package (resolved to https://github.com/golang/sys/).
This package vendors the Go sys subrepository. This is necessary for
Solaris systems. It wasn't obvious until looking through the full
release build logs that this was still being fetched. This pins the
repository to the version used in the v0.1.1 release. Hopefully once Go
has some better dependency management tools, it'll be easier to spot
these things, although it's not a system that's built in testing, so
maybe it would have gone unseen anyway.