Fixed normalizeDirectoryNames for macOS 10.12.
This fix should work, but in the long-run, I think what we're doing here
is the wrong approach and should probably just be scrapped. The problem
is that Apple is going to roll out APFS volumes as in-place conversions
of HFS volumes, and thus the Unicode decomposition will probably be
carried over to these new volumes, so checking for HFS won't be
sufficient. But since APFS doesn't normalize Unicode names, we don't
really want to mess with whatever it has on disk.

This statfs call on EVERY directory probably isn't cheap either.

I think what we should do is just have NFC normalization be an
all-or-nothing thing on macOS systems, probably configurable. For 99.9%
of names, it will make no difference and the conversion won't require
any new string to be allocated because nothing will change. I'd leave it
on by default for now. Then make it configurable once APFS is dominant,
then have it off by default once HFS is a distant memory.

Anyway, we can re-evaluate once APFS rolls out. I've added a TODO.