Switched Entry to use stable marshalling.
This should be less expensive than the full snapshot conversions we were
doing before, and far less error-prone since it's handled by
gogo/protobuf now. It's still not quite as good as we might be able to
do using NamedEntry slices for contents at all times, but it's far less
fragile.

Maybe someday we can switch to using some type of BST or B-tree
representation for the content map of entries, or maybe Go will
eventually support a stable iteration interface for maps, but until that
day this is the best we can reasonably do without some very fragile
code.